% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCC.R
\name{alpha}
\alias{alpha}
\title{Calculate Alpha}
\usage{
alpha(N)
}
\arguments{
\item{N}{Number of standard deviations}
}
\value{
Alpha
}
\description{
This function calculates the percentage of values that lie outside of n standard deviations in a normal distribution. This percentage, often called alpha, is used to set the false positive rate in statistical functions.
}
\details{
\itemize{
\item 3 standard deviations minimizes false positives (1 in 370) at the cost of more missed positives (confirmatory)
\item 2 standard deviations balances false positives (1 in 22) with missed positives (balanced)
\item 1.5 standard deviations accepts more false positives (1 in 7) to avoid missed positives (exploratory)
}
}
