% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCC.R
\name{beta_control_model}
\alias{beta_control_model}
\title{Beta Control DCC Model Specification}
\usage{
beta_control_model(
  inclusion_probabilities = NULL,
  phi_prior = NULL,
  phi_prior_confidence = NULL,
  nmatches = 1
)
}
\arguments{
\item{inclusion_probabilities}{The prior inclusion probabilities of the contemporaneous covariates (optional)}

\item{phi_prior}{The prior concentration (precision) of the beta data model}

\item{phi_prior_confidence}{The confidence of the prior concentration (rate)}

\item{nmatches}{Number of expected contemporaneous covariate matches in the stochastic search (alternative to specifying inclusion probabilities)}
}
\value{
A Beta Control DCC model specification with specified hyperparameters
}
\description{
This function specifies a beta contemporaneous regression data model with stochastic search variable selection (SSVS) for a DCC.
}
