% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCC.R
\name{beta_forecast_model}
\alias{beta_forecast_model}
\title{Beta Forecast DCC Model Specification}
\usage{
beta_forecast_model(phi_prior = NULL, phi_prior_confidence = NULL)
}
\arguments{
\item{phi_prior}{The prior concentration (precision) of the beta data model}

\item{phi_prior_confidence}{The confidence of the prior concentration (rate)}
}
\value{
A Beta Forecast DCC model specification with specified hyperparameters
}
\description{
This function specifies a non-stationary beta data model (i.e., local level model) for a DCC.
}
