% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCC.R
\name{beta_growth_model}
\alias{beta_growth_model}
\title{Beta Growth DCC Model Specification}
\usage{
beta_growth_model(phi_prior = NULL, phi_prior_confidence = NULL)
}
\arguments{
\item{phi_prior}{The prior concentration (precision) of the beta data model}

\item{phi_prior_confidence}{The confidence of the prior concentration (rate)}
}
\value{
A Beta Growth DCC model specification with specified hyperparameters
}
\description{
This function specifies a beta growth data model for a DCC.
}
