% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCC.R
\name{dcc}
\alias{dcc}
\title{Estimate a Dynamic Control Chart}
\usage{
dcc(
  y,
  x,
  intervention_start = NULL,
  ignored = NULL,
  covariates = NULL,
  model = normal_model(),
  nsamples = 10000,
  nburnin = 1000
)
}
\arguments{
\item{y}{Response data}

\item{x}{Subgroup/Time data}

\item{intervention_start}{Time where intervention began}

\item{ignored}{Indices of subgroups/times to ignore when estimating the model}

\item{covariates}{A matrix of covariates for the Response data (optional)}

\item{model}{A DCC model specification}

\item{nsamples}{Number of posterior samples to simulate}

\item{nburnin}{Number of samples to discard (i.e., samples until Bayesian convergence)}
}
\value{
An object of class dcc which is a list with the following components: y, x, intervention_start, ignored, p, full_samples
}
\description{
This function estimates a Dynamic Control Chart using Bayesian statistical methods.
}
