% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCC.R
\name{echart}
\alias{echart}
\title{Estimate an Exchangeability Chart}
\usage{
echart(y, x, stat_fn, nsamples = 10000)
}
\arguments{
\item{y}{Outcome data}

\item{x}{Subgroup data}

\item{stat_fn}{Statistical function for aggregating the outcome data by subgroup}

\item{nsamples}{Number of randomization/permutation samples to simulate}
}
\value{
An object of class echart which is a list with the following components: y, x, stat_fn, full_samples
}
\description{
This function estimates an Exchangeability Chart (E-Chart) using randomization/permutation methods. The E-Chart is a Phase I tool for examining the exchangeability of outcomes across subgroups.
}
