% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCC.R
\name{echart_from_percentages}
\alias{echart_from_percentages}
\title{Estimate an Exchangeability Chart from Subgroup Percentages}
\usage{
echart_from_percentages(
  y,
  x,
  n,
  stat_fn = base::mean,
  ignored = NULL,
  nsamples = 10000
)
}
\arguments{
\item{y}{Percentage data}

\item{x}{Subgroup data}

\item{n}{Size data}

\item{ignored}{Subgroups to ignore when estimating the model}

\item{nsamples}{Number of randomization/permutation samples to simulate}
}
\value{
An object of class echart which is a list with the following components: y, x, stat_fn, ignored, full_samples
}
\description{
This function estimates an Exchangeability Chart (E-Chart) from subgroup percentages and sizes. This is a helper function which generates binary outcome data based on subgroup percentages and sizes and then passes the data to the echart function.
}
