% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCC.R
\name{normal_control_model}
\alias{normal_control_model}
\title{Normal Control DCC Model Specification}
\usage{
normal_control_model(
  inclusion_probabilities = NULL,
  sigma_prior = NULL,
  sigma_prior_confidence = NULL,
  nmatches = 1
)
}
\arguments{
\item{inclusion_probabilities}{The prior inclusion probabilities of the contemporaneous covariates (optional)}

\item{sigma_prior}{The prior standard deviation of the normal data model}

\item{sigma_prior_confidence}{The confidence of the prior standard deviation (rate: sample size)}

\item{nmatches}{Number of expected contemporaneous covariate matches in the stochastic search (alternative to specifying inclusion probabilities)}
}
\value{
A Normal Control DCC model specification with specified hyperparameters
}
\description{
This function specifies a normal contemporaneous regression data model with stochastic search variable selection (SSVS) for a DCC.
}
