% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCC.R
\name{normal_forecast_model}
\alias{normal_forecast_model}
\title{Normal Forecast DCC Model Specification}
\usage{
normal_forecast_model(
  omega_prior = NULL,
  omega_prior_confidence = NULL,
  sigma_prior = NULL,
  sigma_prior_confidence = NULL
)
}
\arguments{
\item{omega_prior}{The prior standard deviation of the innovations (i.e., the change in the non-stationary mean from point to point)}

\item{omega_prior_confidence}{The confidence of the prior standard deviation of the innovations (rate: sample size)}

\item{sigma_prior}{The prior standard deviation of the normal data model}

\item{sigma_prior_confidence}{The confidence of the prior standard deviation (rate: sample size)}
}
\value{
A Normal Forecast DCC model specification with specified hyperparameters
}
\description{
This function specifies a non-stationary normal data model (i.e., local level model) for a DCC.
}
