% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCC.R
\name{normal_growth_model}
\alias{normal_growth_model}
\title{Normal Growth DCC Model Specification}
\usage{
normal_growth_model(
  alpha_prior = NULL,
  alpha_prior_confidence = NULL,
  beta_prior = NULL,
  beta_prior_confidence = NULL,
  sigma_prior = NULL,
  sigma_prior_confidence = NULL
)
}
\arguments{
\item{alpha_prior}{The prior intercept of the normal growth data model}

\item{alpha_prior_confidence}{The confidence of the prior intercept (scale: standard deviation)}

\item{beta_prior}{The prior slope of the normal growth data model}

\item{beta_prior_confidence}{The confidence of the prior slope (scale: standard deviation)}

\item{sigma_prior}{The prior standard deviation of the normal growth data model}

\item{sigma_prior_confidence}{The confidence of the prior standard deviation (rate: sample size)}
}
\value{
A Normal Growth DCC model specification with specified hyperparameters
}
\description{
This function specifies a normal growth data model for a DCC.
}
