% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCC.R
\name{normal_model}
\alias{normal_model}
\title{Normal DCC Model Specification}
\usage{
normal_model(
  alpha_prior_mean = NULL,
  alpha_prior_sd = NULL,
  sigma_prior_mode = NULL,
  sigma_prior_concentration = NULL
)
}
\arguments{
\item{alpha_prior_mean}{The prior mean of the normal data model.}

\item{alpha_prior_sd}{The standard deviation as confidence of the prior mean.}

\item{sigma_prior_mode}{The prior standard deviation of the normal data model.}

\item{sigma_prior_concentration}{The confidence of the prior standard deviation.}
}
\value{
A Normal DCC model specification with specified hyperparameters.
}
\description{
This function specifies a stable normal data model for a DCC.
}
