% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCC.R
\name{param_plot}
\alias{param_plot}
\title{Plot the Parameters of a Dynamic Control Chart}
\usage{
param_plot(dcc, bins = 30, traces = F)
}
\arguments{
\item{dcc}{An estimated DCC object}

\item{traces}{Plot the traces instead of the posterior distributions}
}
\description{
This function plots the posterior distributions for the parameters for a Dynamic Control Chart (DCC).
}
