% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCC.R
\name{plot.echart}
\alias{plot.echart}
\title{Plot an Exchangeability Chart}
\usage{
\method{plot}{echart}(echart, stratify_by = NULL, funnel = F, alpha = 0.002699796)
}
\arguments{
\item{echart}{An estimated E-Chart object}

\item{stratify_by}{Stratification data}

\item{funnel}{Sort the subgroups by N size}

\item{alpha}{False positive rate (Default: 3 standard deviations)}
}
\description{
This function plots an Exchangeability Chart (E-Chart). The plot can be faceted across strata by specifying stratification data.
}
