% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCC.R
\name{plot.ucc}
\alias{plot.ucc}
\title{Plot a Universal Control Chart}
\usage{
\method{plot}{ucc}(ucc, alpha = 0.002699796, funnel = F)
}
\arguments{
\item{ucc}{An estimated Universal Control Chart object.}

\item{alpha}{Statistical significance (false positive rate)}

\item{funnel}{Sort the}
}
\description{
This function plots a Universal Control Chart.
}
