% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCC.R
\name{poisson_model}
\alias{poisson_model}
\title{Poisson DCC Model Specification}
\usage{
poisson_model(lambda_prior_mean = NULL, lambda_prior_concentration = NULL)
}
\arguments{
\item{lambda_prior_mean}{The prior mean of the poisson data model.}

\item{lambda_prior_concentration}{The confidence of the prior mean.}
}
\value{
A Poisson DCC model specification with specified hyperparameters.
}
\description{
This function specifies a stable poisson data model for a DCC.
}
