% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCC.R
\name{beta_model}
\alias{beta_model}
\title{Beta DCC Model Specification}
\usage{
beta_model(
  mu_prior = NULL,
  mu_prior_confidence = NULL,
  phi_prior = NULL,
  phi_prior_confidence = NULL,
  variance_prior = NULL
)
}
\arguments{
\item{mu_prior}{The prior mean of the beta data model}

\item{mu_prior_confidence}{The confidence of the prior mean (rate: sample size)}

\item{phi_prior}{The prior concentration (precision) of the beta data model}

\item{phi_prior_confidence}{The confidence of the prior concentration (rate)}

\item{variance_prior}{The prior variance of the beta data model (alternative to specifying phi prior)}
}
\value{
A Beta DCC model specification with specified hyperparameters
}
\description{
This function specifies a stationary beta data model for a DCC.
}
