% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCC.R
\name{dcc}
\alias{dcc}
\title{Estimate a Dynamic Control Chart}
\usage{
dcc(
  y,
  x,
  intervention_start,
  model = intercept_model(),
  nsamples = 1000,
  nburnin = 1000
)
}
\arguments{
\item{y}{Response data}

\item{x}{Time data}

\item{intervention_start}{The index in the data where the intervention began.}

\item{model}{A DCC model specification}

\item{nsamples}{The number of counterfactual samples to simulate.}

\item{nburnin}{The number of samples to discard (i.e., samples until Bayesian convergence).}
}
\value{
An object of class dcc which is a list with the following components: y, x, intervention_start, p, full_samples.
}
\description{
This function estimates a model from baseline data and simulates counterfactual data for future time points.
}
