% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCC.R
\name{dcc}
\alias{dcc}
\title{Estimate a Dynamic Control Chart}
\usage{
dcc(
  y,
  x,
  intervention_start = NULL,
  ignored = NULL,
  covariates = NULL,
  model = normal_model(),
  nsamples = 10000,
  nburnin = 1000,
  param_diagnostic = T
)
}
\arguments{
\item{y}{Outcome data}

\item{x}{Time series data}

\item{intervention_start}{Index when change began}

\item{ignored}{Indices of data points to ignore when estimating the model}

\item{covariates}{A matrix of contemporaneous covariates for the outcome data (optional)}

\item{model}{A DCC model specification}

\item{nsamples}{Number of posterior samples to simulate}

\item{nburnin}{Number of posterior samples to discard (i.e., samples until Bayesian convergence)}
}
\value{
An object of class dcc which is a list with the following components: y, x, intervention_start, ignored, covariates, p, full_samples
}
\description{
This function estimates a Dynamic Control Chart (DCC) using Bayesian statistical methods. The DCC is a tool for substantiating whether a change resulted in an improved outcome. The DCC can also be used for forecasting an outcome during goal setting.
}
