% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCC.R
\name{plot.dcc}
\alias{plot.dcc}
\title{Plot a Dynamic Control Chart}
\usage{
\method{plot}{dcc}(dcc)
}
\arguments{
\item{dcc}{An estimated Dynamic Control Chart object.}
}
\description{
This function plots a Dynamic Control Chart.
}
