% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCC.R
\name{plot.dcc}
\alias{plot.dcc}
\title{Plot a Dynamic Control Chart}
\usage{
\method{plot}{dcc}(dcc, alpha = 0.002699796)
}
\arguments{
\item{dcc}{An estimated Dynamic Control Chart object.}

\item{alpha}{Statistical significance (false positive rate)}
}
\description{
This function plots a Dynamic Control Chart.
}
